add soluion of Drinks problem

+++ Drinks.cpp
@@ -2,10 +2,10 @@
 using namespace std;
 
 int main() {
-    float n;
+    int n;
     cin >> n;
-    float temp = n;
-    float sum = 0;
+    int temp = n;
+    int sum = 0;
 
     while(n--) {
         int num;
@@ -13,5 +13,5 @@ int main() {
 
         sum += num;
     }
-    cout << (float) sum/(temp);
+    cout << (double)sum/temp;
 }
